node/bindnode: add native support for cid.Cid
That is, support the Go type go-cid.Cid to store IPLD links,
not just the ipld.Link interface.

In AssignLink, this requires unboxing the cidlink.Link
when the target Go type is a go-cid.Cid.
Similarly, in AsLink, we need to box a go-cid.Cid as a cidlink.Link.

As a means to test this feature,
we start adding the first table-driven Prototype tests.
These are specific to Bindnode, as they include Go types.

We add a Scalars test case with what already worked,
and a Links test case which uses both go-cid.Cid and ipld.Link.

Note that we needed to add the Link type to dmt.Compile,
as it's often defined as part of the prelude but I had forgotten it.
We also make the "schemafree link emission" error include a type,
as I ran into the error and had to printf-debug what I did wrong.