start using Rod's schema tests from ipld/ipld
This is via a new test in schema/dsl,
which has full access to both the DSL parsing and the DMT compiling.
Note that this pulls in a yaml dependency,
but the dependency is only required for running tests,
and it will probably be replaced soon in favor of Eric's testmark.

We start with just under half of the test yml files being enabled,
as those are the simpler cases that our stack already supports.

That said, we needed to improve support for representation strategies
across the three packages involved.
In paraticular, we were missing support for map renames,
enums having no explicit string representation key,
and kinded unions holding integers in bindnode.

Finally, add initial support for "&" links in dsl and dmt,
even though we have no tests for those yet.
That's because Rod's tests with links use "Any",
which the schema package does not support yet.

Updates #188.