node/bindnode: add support for all scalars behind kinded unions
AssignBytes also didn't call the finish func - oops.
We caught this because AssignBytes on a kinded union wasn't actually
storing anything, since that happens via a finish func on kinded unions.

Also prevent a panic if an union has an invalid state.
That is, when none of its pointer fields are non-nil.
While at it, panic if any of its fields are not pointers.

The changes to improve kinded union support are rather repetitive.