storage: add 'Has' feature.
I debated for a while if there's any way to avoid doing interfaces that
refer to other interfaces here, because I usually find that's a smell,
but any idea I can think of to try to avoid it seems worse (e.g.
would threaten to spawn a 'Has(R)' *and* a 'Has2(W)' function on the
package scope, due to golang's lack of polymorphic parameters, etc).