node/bindnode: start verifying schema compatibility
That is, if the user calls Prototype or Wrap with an IPLD Schema and Go
type which are not compatible, we fail immediately with a reasonably
useful error message, rather than panicking or failing in weird ways at
a later time.

Note that, much like bindnode itself, verification doesn't support all
edge cases of the schema type system just yet.

Fixes #269.