node/bindnode: export ProduceGoTypes
Now that the DSL parser and DMT compiler are mostly done,
there's no reason to hide this API anymore.
Remove the TODO as well, because the feature will be exposed via
ipldtool - so there's no point in yet another binary.

The existing API seems fine as it is. Finally, add docs.

While here, re-running "go generate ./..." produced small changes,
as gendemo was out of date.