storage: update fsstore defaults to be the same as flatfs as used by default configuration of go-ipfs.
Include links to where the relevant ipfs code can be found.
(It's not all in very obvious places, and not all in the go-ipfs repo.)

Extracted some of the b32 encoder to a package function...
but, although this does avoid some repeated initialization,
it was all CPU stuff.  Benchmarks indicate that it made no real
difference.  So it's mostly a stylistic change.