storage: switch storage.StreamingReadableStorage.GetStream to return an io.ReadCloser instead of just io.Reader.
This seems more likely to encourge good code that doesn't leak
resources than just admonishing people to check via a comment.

On the downside, sometimes this requires implementations to wrap stuff
in no-op closers.  The trade is probably worth it.

The code in LinkSystem which uses these things still doesn't change.
It's still using the `linking.Block*Opener` functions, which still
return just io.Reader.  And we're considerably less free to change
those, because there are already quite a few implementations of those
types in the wild, and we don't want to break them.