node/bindnode: allow binding cidlink.Link to links
We allowed the low-level go-cid.Cid as well as the generic ipld.Link,
but not the type that implements ipld.Link via a go-cid.Cid.
Allow all three to be consistent and nice towards the user.

The get/set code already handles the case correctly;
it's just verifyCompatibility that didn't account for it.

While here, make "kind mismatch" messages more verbose,
as otherwise they may be confusing to the user.

Fixes #305.