schema: avoid alloc and copy in Struct and Enum methods
Querying the fields and members of structs and enums is common.
I get the intent of trying to keep them read-only,
but the current mechanism is just too expensive.

	name                                 old time/op    new time/op    delta
	MessageEncodingRoundtrip/DagCbor-16    21.4µs ± 1%    16.9µs ± 0%  -20.66%  (p=0.029 n=4+4)

	name                                 old alloc/op   new alloc/op   delta
	MessageEncodingRoundtrip/DagCbor-16    32.8kB ± 0%    23.2kB ± 0%  -29.28%  (p=0.029 n=4+4)

	name                                 old allocs/op  new allocs/op  delta
	MessageEncodingRoundtrip/DagCbor-16       451 ± 0%       396 ± 0%  -12.20%  (p=0.029 n=4+4)

An alternative could be to provide a slice,
and then we'd append into it and reuse its space.
However, note that that would still copy memory.

An alternative could be to provide iterator-like methods, such as:

	Length() int
	FieldAt(int) StructField

However, that would get relatively clunky, and doesn't seem particularly
worth it right now.

Ultimately, we don't have read-only slices, so we shouldn't try to
invent our own. The fact that we return the direct slice does allow
replacing elements, but the user can't append or re-slice, so that
should be enough signal to most Go developers.