schema: remove one review note; add another.
The idea of union typekind always acting like a map seemed like it
might be a bit kludgey when I first came up with it, but has actually
stood up well and has not caused any other design principle conflicts;
the hesitancy in the comment there no long seems valid.

On the other hand, noticing the behavior of the unit typekind...
seeing those as a boolean seems like a choice that may have been made
in somewhat whimsy (and probably while thinking overmuch of conventions
of maps-for-sets in the wild), and I would probably like to review it.