node/bindnode: cover even more edge case panics
On the read side, we now also cover the misuse of LookupByString,
LookupByIndex, Length, and Prototype. That caught a panic in
LookupByIndex on tuples.

On the write side, we now also cover the misuse of AssembleKey and
AssembleEntry, which could panic with structs.

Finally, we add enums and kinded unions as test cases.
Enums already worked fine with all the edge cases,
but kinded unions did not support ListIterator at the repr level
and could panic when assigning mismatching kinds.

There's still work to do to polish the sharp edges of the bindnode
package, but this should have us covered in terms of ensuring the
commonly used methods don't panic. Further testing will happen as part
of the fuzzing effort in #357.

Fixes #342.