node/bindnode: make AssignNode work at the repr level
The repr-level AssignNode just called the type-level AssignNode.
This worked for cases where the repr level was identical,
such as a struct with map repr, so we hadn't caught the mistake yet.

The problem is that forwarding the call from the repr level to the type
level would then call the Assembler methods on the type level.
In the case of our test, that would fail, as the renames would not be
taken into account properly.

The fix is to make the two methods call the assembler methods on their
own types, be it type level or repr level.
Luckily, we don't need to copy-paste the code or move it around,
because datamodel.Copy has luckily existed since v0.14.0.

Thanks to Rod Vagg for writing the reproducer in the form of a test.

Fixes #365.