node/bindnode: support pointers to datamodel.Node to bind with Any
We missed calling nonPtrVal in a few places when reading from Go values.
The added test case panicked due to the missing call in _structIterator.

Note that the added test case only covers one of the nonPtrVal calls,
but I reckon this is enough for now, given the consistency in the code.
The other modified lines are still covered by existing tests,
just not with the particular pointer interaction.

Rod ran into the same panic while developing go-graphsync,
so this should also help unblock their progress.