Allow feature detection for the `io.ReaderAt` interface on Bytes kind nodes
We can handle large bytes by allowing Bytes kinded nodes to optionally support
the `ReadAt` method thereby implementing the ReaderAt interface.

This is proposed over a distinct the reader / seeker implementation because the ReaderAt interface is stateless - it does not need the addition of a position cursor to be tracked in the node between calls, while still allowing for partial reads of the node contents.