node/bindnode: start fuzzing with schema/dmt and codec/dagcbor
The fuzzing function has two []byte inputs:
the first is an IPLD schema-schema in dag-cbor form,
and the second is an IPLD node in dag-cbor form.

We chose dag-cbor because it's concise, which helps fuzzing,
but also because it is widely used as an IPLD codec.

This lets us fuzz most of the bindnode and dagcbor packages.
For the time being, panics from schemadmt.Compile are ignored,
because both the schema and schemadmt packages are prone to them.
We will eventually fix them, but for the time being, they are not a
priority as they can't possibly affect users like GraphSync.

Finally, add and fix two crashers.

First, a schema struct with a field named "0". We would attempt to infer
a Go struct type with a field named "0", which is invalid in Go, and
would cause a reflect panic.

Second, a schema enum with int repr being decoded at the repr level from
a string. Instead of returning an "invalid kind" error, we would happily
assign the string to the Go value, which would then make the re-encode
panic in a weird way.

Updates #357.