codec/dagcbor: add DecodeOptions.ExperimentalDeterminism
This is simialr to Rod Vagg's EncodeOptions.MapSortMode,
which taught the encoder to canonically sort map keys as per the spec,
but now we're doing the decode side.

We add a single "strictness" knob, as it needs to deal with multiple
bits like integers and floats, and not just map key order.

Update the fuzzer to be aware of this setting.