Readme updates!
Update the description of change cycles a bit.
We've become significantly more careful since this was last penned.

Described the relationship of unixfs code to this repo much better,
and point more directly at implementations.  Attached some "why".
(The "why" might be a bit much for the front page README, in the long
run -- but for now, I think it may be useful, given how often
questions specific to that topic and subsystem come up.)

Point directly to website pages that mention other golang libraries.