Fix hash mismatch error on matching link pointer
On load, the default IPLD Link System checks the hash of loaded data
against the given hash to assure that they match. The actual hash is
calculated using the given link's `LinkPrototype.BuildLink`, and then is
directly compared with the given link via `==`.

Because the comparison is done by value, if the given link is a pointer
the comparison fails which will result in false-positive hash mismatch
error.

To address this issue, instead of directly comparing objects using `==`
check that their `Binary` value, i.e. their densest possible encoding,
is equal. Add tests that assert the issue is fixed and works with
existing link representations.