Export schema/dmt.TypeSystem.
This is both necessary and sufficient to make it feasible for someone
to easily manipulate schema DMT data as IPLD, using the IPLD libraries.
That's a fairly obvious win.

(If I had my way about the language of golang, I'd be able to do this
while also keeping it strictly immutable.  That's... not how it works,
unfortunately.  But as it's normal in golang to have package-scoped
vars with semantic rules about no-touchy and no actual
compiler-enforced support, then, we might as well not deny ourselves
useful things by wishing for support for stricter rules that just
doen't exist.)