Add several pieces of docs to schema/dmt.
Package docs are important to have as overviews,
and we were previously missing any.

Added a few notes to any future hackers, as well.

Removed the "export" file that was only used by the
generator gadget.  I don't believe it worked anyway.
It wasn't exercised in CI because of the build tags.
And since it was first written, we've relaxed on
what we allow to be exported anyway, and the field
that file contained a build-tag-fenced accessor for is
simply available anyway.  So, I updated the codegen
gadget to just use that directly.  No sweat.